Refactor code: Remove unused functions and imports, and update function names
