+++ utils/test/testUtils.js
@@ -70,6 +70,43 @@ function compareImages(imageDataURL, baseline, outputName) {
   });
 }
 
+/**
+ * Adds two concentric circles to each axial slice of the demo segmentation.
+ */
+function createMockEllipsoidStackSegmentation({
+  imageIds,
+  segmentationImageIds,
+  cornerstone,
+}) {
+  const { metaData, cache } = cornerstone;
+  const { rows, columns } = metaData.get('imagePlaneModule', imageIds[0]);
+  const dimensions = [columns, rows, imageIds.length];
+
+  const center = [dimensions[0] / 2, dimensions[1] / 2, dimensions[2] / 2];
+  const outerRadius = 64;
+  const innerRadius = 32;
+  for (let z = 0; z < dimensions[2]; z++) {
+    let voxelIndex = 0;
+    const image = cache.getImage(segmentationImageIds[z]);
+    const scalarData = image.getPixelData();
+    for (let y = 0; y < dimensions[1]; y++) {
+      for (let x = 0; x < dimensions[0]; x++) {
+        const distanceFromCenter = Math.sqrt(
+          (x - center[0]) * (x - center[0]) +
+            (y - center[1]) * (y - center[1]) +
+            (z - center[2]) * (z - center[2])
+        );
+        if (distanceFromCenter < innerRadius) {
+          scalarData[voxelIndex] = 1;
+        } else if (distanceFromCenter < outerRadius) {
+          scalarData[voxelIndex] = 2;
+        }
+        voxelIndex++;
+      }
+    }
+  }
+}
+
 export {
   fakeImageLoader,
   fakeMetaDataProvider,
@@ -78,4 +115,5 @@ export {
   createNormalizedMouseEvent,
   // utils
   colors,
+  createMockEllipsoidStackSegmentation,
 };
